requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
langchain-groq
langchain-pinecone
pinecone
python-dotenv
pypdf
python-multipart